requirements-agent.txt
# Core
playwright>=1.40.0
cryptography>=41.0.0

# Telegram agents
telethon>=1.34.0

# EVM crypto payments
web3>=6.0.0

# Optional: BTC payments
# bitcoinlib>=0.6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
